+++ pages/_app.tsx
@@ -7,7 +7,7 @@ import Link from 'next/link'
 import CardContent from '@material-ui/core/CardContent'
 import Grid from '@material-ui/core/Grid'
 
-import { useApollo } from '../src/lib/apolloClient'
+import { useApollo } from 'src/lib/apolloClient'
 
 import 'src/components/App/styles/styles.less'