+++ changelog.md
@@ -1,8 +1,23 @@
 # 1.0.0
 
+===============================================
+
 - Added typescript
 - Added apollo-client
 - Added jest tests
 - Added eslint
 - Added prettier
 - Added graphql schema and types generator
+
+# 1.1.0
+
+===============================================
+
+- Code refactoring
+- Added apollo tests
+
+# 1.1.1
+
+===============================================
+
+- Fix API endpoint for vercel.com deployment